Revert "Phase 3: Java Console Programming with JDBC (#11)"
This reverts commit cec439a6e421425b82bdc1ddc2c5d8d3ed9b3f11.
