--- .gitignore
@@ -1,3 +1,0 @@
-.DS_Store
-/.DS_Store
-*/.DS_Store
