+++ phase-2/sql/eng/bus_schema.sql
@@ -91,7 +91,7 @@ CREATE TABLE RESERVATION(
     RAid CHAR(9) NOT NULL,
     RSid CHAR(2) NOT NULL,
     RTid NUMBER NOT NULL,
-    PRIMARY KEY (RSid, RTid),
+    PRIMARY KEY (RAid, RSid, RTid),
     FOREIGN KEY (RAid) REFERENCES ACCOUNT(Aid),
     FOREIGN KEY (RSid) REFERENCES SEAT(Sid),
     FOREIGN KEY (RTid) REFERENCES TIMETABLE(Tid)
